requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
kombu==5.2.4
multidict==6.0.4
openai==0.27.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
python-dotenv==1.0.0
python-telegram-bot==20.1
pytz==2022.7.1
redis==4.5.2
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.8.2
pydub~=0.25.1